requirements.txt
tensorflow>=2.12.0
opencv-python
numpy
pandas
matplotlib
scikit-learn
Pillow
Flask
